Feat: board 댓글 달기 및 삭제, 단일 글 조회 수정

+++ src/main/java/POT/DuoBloom/board/controller/BoardController.java
@@ -1,4 +1,4 @@
-package POT.DuoBloom.board;
+package POT.DuoBloom.board.controller;
 
 import POT.DuoBloom.board.dto.BoardRequestDto;
 import POT.DuoBloom.board.dto.BoardResponseDto;
@@ -11,7 +11,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -21,7 +20,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @CrossOrigin(origins = "http://localhost:3000")
 @RestController
-@RequestMapping("/api/feeds/board")
+@RequestMapping("/api/feeds/boards")
 public class BoardController {
 
     private final BoardService boardService;
@@ -31,14 +30,15 @@ public class BoardController {
     @PostMapping
     public ResponseEntity<BoardResponseDto> createBoard(@RequestBody BoardRequestDto boardRequestDTO, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
-
         if (userId == null) {
-            return ResponseEntity.status(401).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
+        if (!boardService.canAccessBoard(user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         Board board = boardService.createBoard(user, boardRequestDTO.getTitle(), boardRequestDTO.getContent());
-        BoardResponseDto responseDTO = new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt());
+        BoardResponseDto responseDTO = new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt(), null);
         return ResponseEntity.ok(responseDTO);
     }
 
@@ -47,7 +47,7 @@ public class BoardController {
     public ResponseEntity<List<BoardResponseDto>> getAllBoards() {
         List<Board> boards = boardService.getAllBoards();
         List<BoardResponseDto> responseDTOs = boards.stream()
-                .map(board -> new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt()))
+                .map(board -> new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt(), null))
                 .collect(Collectors.toList());
         return ResponseEntity.ok(responseDTOs);
     }
@@ -56,21 +56,14 @@ public class BoardController {
     @GetMapping("/{boardId}")
     public ResponseEntity<BoardResponseDto> getBoardById(@PathVariable Integer boardId, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
-
         if (userId == null) {
-            return ResponseEntity.status(401).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
-
-        // 접근 권한 확인
         if (!boardService.canAccessBoard(user)) {
-            throw new IllegalStateException("커플인 경우에만 커뮤니티 글을 조회할 수 있습니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
-
-        Board board = boardService.getBoardById(boardId)
-                .orElseThrow(() -> new IllegalStateException("해당 글이 존재하지 않습니다."));
-        BoardResponseDto responseDTO = new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt());
+        BoardResponseDto responseDTO = boardService.getBoardDetailsById(boardId);
         return ResponseEntity.ok(responseDTO);
     }
 
@@ -78,14 +71,15 @@ public class BoardController {
     @PutMapping("/{boardId}")
     public ResponseEntity<BoardResponseDto> updateBoard(@PathVariable Integer boardId, @RequestBody BoardRequestDto boardRequestDTO, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
-
         if (userId == null) {
-            return ResponseEntity.status(401).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
+        if (!boardService.canAccessBoard(user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         Board board = boardService.updateBoard(user, boardId, boardRequestDTO.getTitle(), boardRequestDTO.getContent());
-        BoardResponseDto responseDTO = new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt());
+        BoardResponseDto responseDTO = new BoardResponseDto(board.getBoardId(), board.getTitle(), board.getContent(), board.getUpdatedAt(), null);
         return ResponseEntity.ok(responseDTO);
     }
 
@@ -93,41 +87,38 @@ public class BoardController {
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Void> deleteBoard(@PathVariable Integer boardId, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
-
         if (userId == null) {
-            return ResponseEntity.status(401).build();
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
+        if (!boardService.canAccessBoard(user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
         boardService.deleteBoard(user, boardId);
         return ResponseEntity.noContent().build();
     }
 
     // 좋아요 추가
-    @PostMapping("/{boardId}")
+    @PostMapping("/{boardId}/like")
     public ResponseEntity<Void> likeBoard(@PathVariable Integer boardId, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
         if (userId == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
         boardService.likeBoard(user, boardId);
         return ResponseEntity.ok().build();
     }
 
     // 좋아요 취소
-    @PatchMapping("/{boardId}")
+    @PatchMapping("/{boardId}/unlike")
     public ResponseEntity<Void> unlikeBoard(@PathVariable Integer boardId, HttpSession session) {
         Long userId = (Long) session.getAttribute("userId");
         if (userId == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
-
         User user = userService.findById(userId);
         boardService.unlikeBoard(user, boardId);
         return ResponseEntity.ok().build();
     }
-
-
 }
